commitlog: destroy segment allocation future only once
commit 8a745593a217dea437d3e697f22ae4953b44d624 fixed a race
condition that caused two tasks to do pop_eventually simultaneously.
However it introduced some kind of a "double free" for that future
that caused a similar race.
This change amends to the previous one in order to avoid that by
making sure only the first allocator that resumes will free this future.

Refs #10020
Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>