test.py: make case cache global
test.py runs each unit test's test case in a separate process.
The list of test cases is built at start, by running --list-cases
for each unit test. The output is cached, so that if one uses --repeat
option, we don't list the cases again and again.

The cache, however, was only useful for --repeat, because it was only
caching the last tests' output, not all tests output, so if I, for example,
run tests like:

./test.py foo bar foo

.. the cache was unused. Make the cache global which simplifies its
logic and makes it work in more cases.