sstables: validate originating host id
Add an additional sstable validation step to check that originating
host id matches the local host id.

This is only done for ME-and-up sstables, which do not come from
upload/, and when the local host id is known.

When local host id is unknown, check that the sstable belongs to a
system keyspace, i.e. whether it is plausible that Scylla is still
booting up and hasn't loaded/generated the local host id yet.

Signed-off-by: Michael Livshin <michael.livshin@scylladb.com>