storage_service: De-bloat waiting logs
First thing is that logging can be done with logger methods,
not with set_mode() because the mode is already set at this
place.

Second thing is that pre-update_pending_ranges logs are excessive,
as the update_pending_ranges logs its progress itself.

Third is that post-logging is also exsessive -- there are more
logs after those lines.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>